refactor stackscript logic
