+++ src/features/StackScripts/SelectStackScriptPanel/index.ts
@@ -1,3 +1,2 @@
 import SelectStackScriptPanel from './SelectStackScriptPanel';
-// export { CommunityStackScripts, LinodeStackScripts, MyStackScripts };
 export default SelectStackScriptPanel;
