+++ src/features/linodes/LinodesDetail/LinodeRebuild/LinodeRebuild.tsx
@@ -29,10 +29,14 @@ interface ContextProps {
 }
 type CombinedProps = WithStyles<ClassNames> & ContextProps;
 
-type MODES = 'fromImage' | 'fromStackScript';
+type MODES =
+  | 'fromImage'
+  | 'fromCommunityStackScript'
+  | 'fromAccountStackScript';
 const options = [
   { value: 'fromImage', label: 'From Image' },
-  { value: 'fromStackScript', label: 'From StackScript' }
+  { value: 'fromCommunityStackScript', label: 'From Community StackScript' },
+  { value: 'fromAccountStackScript', label: 'From Account StackScript' }
 ];
 
 const LinodeRebuild: React.StatelessComponent<CombinedProps> = props => {
@@ -66,7 +70,12 @@ const LinodeRebuild: React.StatelessComponent<CombinedProps> = props => {
         />
       </Paper>
       {mode === 'fromImage' && <RebuildFromImage />}
-      {mode === 'fromStackScript' && <RebuildFromStackScript />}
+      {mode === 'fromCommunityStackScript' && (
+        <RebuildFromStackScript type="community" />
+      )}
+      {mode === 'fromAccountStackScript' && (
+        <RebuildFromStackScript type="account" />
+      )}
     </React.Fragment>
   );
 };
